backend/requirements.txt
# Django Core
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.2
drf-spectacular==0.29.0

# Database
psycopg2-binary==2.9.10
dj-database-url==2.3.0

# Environment
python-dotenv==1.1.1
python-decouple==3.8

# Static Files
whitenoise==6.11.0

# Channels (WebSocket)
channels==4.3.1
channels-redis==4.3.0
daphne==4.2.1

# Task Queue
celery==5.4.0
django-celery-beat==2.7.0
redis==5.2.0

# AI/ML
google-generativeai==0.8.5

# External APIs
requests==2.32.5
python-telegram-bot==22.5

# Utilities
Pillow==12.0.0

# Production Server
gunicorn==23.0.0
